refactor(ZNTA-1865): Rename ProjectVersionDisplay to project-version-displays
Was exporting ProjectVersionVertical as a Module but naming convention
was that of Component
