+++ server/zanata-frontend/src/frontend/app/containers/ProjectVersion/project-version-displays.js
